Доделал базовый функционал классов, начата реализация системы уровней

+++ src/main/java/org/hlebushek/playerClasses/model/RunnerConfig.java
@@ -1,4 +1,4 @@
 package org.hlebushek.playerClasses.model;
 
-public record RunnerConfig(int speed_level, int fall_damage_multiplier) {
+public record RunnerConfig(int speed_level, double fall_damage_multiplier) {
 }
